Instruments added and updated (HI, compass, alt)
Added a new heading indicator, based on the generic FG model: it features a new
rotating card with larger numbers and letters, as well as a fixed jet shape
insted of the Cessna shape.

Added a magnetic compass as well, taken from the Yak-18T and unmodified.

The altimeter has a new, cleaner dial, recreated from the original one.

The vacuum system was slightly modified to allow the heading indicator to work
properly. In particular, the "scale" parameter was increased to 20, in order
to compensate the low engine rpm of the model. The systems file was copied from
the Generic directory and imported into this model. The resulting redundant
reference to the electrical systems was corrected and left in the systems file
only.

The position of a few instruments was slightly modified (alignment, etc).